requirements.txt
aiohttp==3.11.9
discord.py==2.4.0
matplotlib==3.9.3
numpy==2.2.0
Pillow==11.0.0
